assets: vendorize cosign binaries
Use the newly introduced assets library to bundle the cosign binaries,
required to check the signatures and prepare container images on
airgapped environments.

The reason we vendorize cosign rather than depending on it is that it's
not packaged for Fedora, and only recent versions of Debian ship with
it (at the time of writing, only present in testing). For Windows and
macOS, we package it so we can depend on them.

Fixes #1062