Simplify our Debian instructions
Simplify our Debian instructions using a single `gpg` command for all
distros.  Previously, we used `sq` for certain distros, because we
didn't have a nice way to get our GPG key from the keyservers in the
format that Debian expects [1]. Since then, we've found that you can
prepend the `--keyring` argument with `gnupg-ring:` [2,3] to make GPG
use the desired format. We can also fix the lack of `dirmngr` by using
a random `--homedir`, as mentioned in the GPG docs [4].

[1] https://github.com/freedomofpress/dangerzone/issues/1052#issuecomment-2594065503
[2] https://www.gnupg.org/gph/en/manual/r1574.html
[3] https://stackoverflow.com/a/72278246
[4] https://www.gnupg.org/documentation/manuals/gnupg/Ephemeral-home-directories.html

Refs #1052
Closes #1158