ci: Use a regular user for the check_repos test
Instead of running the Debian/Ubuntu instructions in a container that
has root, and cannot run `sudo`, use a regular user instead. This is
important because we've noticed that installing as root does not catch
several corner cases.

We are adding a Dockerfile for that purpose on the repo, so that it's
easy to make this test in our CI, but also locally as well.