feat: Detect when new container images are out
As part of this change, the release-check logic, which was living
inside the gui, has been separated from it and now lives in
`updater/releases.py`.

Reports have been extended to mention if new container images are out.

In order to avoid making any HTTP requests during runtime (except when
the upgrade process happens), we read the container registry and store
the results in the settings, which will be read by the `installer.py`
module when needed.