Reliably pass seccomp profile to container runtime
Pass the seccomp profile to the container runtime, and make sure that it
works across all platforms/runtime combinations.

With Podman on Linux and Docker Desktop on Windows/macOS, this was not
an issue, but with the introduction of Podman support, it now is. We now
need to copy the seccomp profile to a place that Podman Desktop can
read, and translate that path to the WSL2-equivalent (Windows-only).

Refs #1127
Closes #1187
Closes #1191