docs: Simplify Podman Desktop instructions
Remove the need to create a custom `containers.conf` file, now that we
have a way to pass the seccomp policy to the Podman machine correctly.
Also, do not ask the user to start/stop a Podman machine, as it may be a
bit more involved. Instead, ask them only to verify that they are
running in rootless mode, and provide instructions to do so.

Closes #1127