Make seccomp allow unknown syscalls for podman versions < 3.3
In order to make it work, the used seccomp policy is the same one as the
one shipped by gVisor, except we allow unknown syscalls rather than
denying them.

That fixes an issue that happens in Debian Bullseye where the "mseal"
system call is being denied with ENOPERM rather than the expected
ENOSYS, making the conversions fail.

Fixes #1201