Never install numpy 2.0.0 in Python 3.13 envs
We have encountered some issues when installing numpy 2.0.0 in
Python 3.13 environments [1]. In a previous commit, we circumvented
those by installing numpy 2.0.0 only on Linux, and using newer numpy
versions in Windows / macOS.

It turns out though that installing numpy 2.0.0 in Linux is not ideal,
especially in Python 3.13 environments. The reason is that the
installation process actually builds numpy from source. When this
happens in a VM, it can be prohibitively slow, and take over 10 minutes
to complete.

Therefore, it's best to install the oldest version only in Python 3.9
environments, and use the newer wheels for the rest.

[1] https://github.com/freedomofpress/dangerzone/issues/1206