Create git archive from HEAD
When we use `doit` to build the artifacts, we may do it at a time where
the final tag has not been pushed yet, and instead an RC tag exists.
Unfortunately, `doit` fails in this case, because it expects a proper
release tag (e.g., v0.9.1).

Relax the archival function by archiving HEAD instead. After all, it's
up to the developer to create the final artifacts from a clean
environment.