Use archived debian bullseye backports for testing
Debian bullseye stopped providing backports for this LTS release,
breaking our testing scripts. This makes use of the archived backports,
which is not an ideal situation but has the benefit of making us having
tests for this platform.

Other solutions are currently being discussed, so this could be seen as
a temporary fix.

Fixes #1213