podman: Vendor PodmanCommand from podman-py repo
In the interest of giving back to the Podman community, we had sent a PR
to the containers/podman-py repo that allows Python users to issue
`podman machine` commands from Windows/macOS platforms in a more
Pythonic way.

Because the fate of this PR is not known as of yet, we have decided to
vendor the code we've written for this PR *as is* under
`dangerzone/podman/command`. We have also copied the `podman.errors`
module, since it's used by our PR.

Refs containers/podman-py#558
Refs #1172