podman: Add CLI command to manage Podman machines
Add a CLI command (`dangerzone-machine`) to handle Podman machines for
Dangerzone. This command can use the vendored Podman binary and pass it
a specially crafted containers.conf file. It's mostly a convenience
wrapper, in case a user needs to manually stop or start a Podman
machine. It's also pretty useful for devs to debug a Podman machine
issue locally.

Refs #1171