gui: Run startup tasks in background thread
Replace the various threads we had for checking for updates and
installing container images, with a single thread that runs in the
background and performs some startup tasks.

Also, connect each event from these startup tasks with handler of the
main window, in order to perform an action for every conceivable task
transition.

Fixes #118
Fixes #1045
Fixes #1145
Fixes #1171
Fixes #1184
Closes #253
Closes #470
Closes #1132