Speed up container startup times on Wndows/macOS
Now that we control the Podman machine on Windows and macOS
environments, add an option to use FUSE overlayfs in these environments.
We prefer FUSE overlayfs instead of native overlayfs, because the former
does not need to `chown` the IDs of the files during startup, and is
thus much faster.

Note that we purposefully don't use FUSE overlayfs in two cases a) on
Linux and b) when external Podman is used. The reason is that in these
cases, Podman is not under our control, which may create issues with any
pre-existing storage driver.

Refs #449