docs: Update doc on releasing a Dangerzone image
Update the previous instructions on releasing a Dangerzone image for
the Independent Container Updates feature. The new instructions refer to
the GHCR signer project, attesting provenances, and add some extra
cautionary steps for the maintainer. Also, the document is now part of
the release process, but we state that it can run independently as well.

Closes #1162