Properly strip `v` prefix from version
We want to strip the `v` prefix from the string returned by
`git describe`, but we have to take into account that `git describe`
does not always return a string like `v0.9.0-...`. It may also return a
string like `hotfix-0.9.1-123-gf457d48d`, whose first character we must
not strip.

A better way is to strip just the `v` character from the start with
`sed`.