gui: Style the status bar AND the message
We can't style the QStatusBar and the QLabel with a single style,
because somehow the QLabel is not encompassed within the status bar, at
least for styling purposes. For now, we explicitly pass the style
property we want to the message as well.

Also, remove the grip icon from the bottom right of the status bar,
since it's not visually aesthetic.