Relax the way we attest provenenances
Remove some checks for the Git commit and branch that we were doing
while attesting provenances, since it's easy for an image to have
multiple provenances attached to it, from different branches and Git
commits. Instead, we'll now just verify that the image was built from
our Git repo and a specific workflow. The Git commit verification should
be covered by our reproducibility checks.

Closes #1270