Pin the Rekor public key
Use the bundled Rekor public key in `share/rekor.pub` when verifying
images with Cosign. This way, Dangerzone can operate in cases where
there is no network connectivity. The caveat is that once Sigstore
rotates the Rekor public key, we'll need to publish a new Dangerzone
version with the new key, in order for Dangerzone to verify **new**
container images. In practice, this hasn't happened yet, but even if it
does, we still allow users to override the `SIGSTORE_REKOR_PUBLIC_KEY`
envvar, and set the updated key.

Refs #1280